wip with nss_wrapper

+++ entrypoint.sh
@@ -1,31 +1,39 @@
 #!/bin/bash
+
+# Create passwd file for nss_wrapper
+source /etc/loolwsd/generate_container_user
+
 set -e
 
 # TODO replace this with a cleaner implementation
 if test "${DONT_GEN_SSL_CERT-set}" == set; then
-  # Generate new SSL certificate instead of using the default
-  mkdir -p /home/lool/ssl/
-  cd /home/lool/ssl/
-  mkdir -p certs/ca
 
-  openssl genrsa -out certs/ca/root.key.pem 2048
-  openssl req -x509 -new -nodes -key certs/ca/root.key.pem -days 9131 -out certs/ca/root.crt.pem -subj "/C=DE/ST=BW/L=Stuttgart/O=Dummy Authority/CN=Dummy Authority"
+  if [ ! -f /home/lool/ssl/certs/ca/root.key.pem ]; then
+    # Generate new SSL certificate instead of using the default
+    mkdir -p /home/lool/ssl/
+    cd /home/lool/ssl/
+    mkdir -p certs/ca
 
-  mkdir -p certs/{servers,tmp}
-  mkdir -p "certs/servers/localhost"
+    openssl genrsa -out certs/ca/root.key.pem 2048
+    openssl req -x509 -new -nodes -key certs/ca/root.key.pem -days 9131 -out certs/ca/root.crt.pem -subj "/C=DE/ST=BW/L=Stuttgart/O=Dummy Authority/CN=Dummy Authority"
 
-  openssl genrsa -out "certs/servers/localhost/privkey.pem" 2048 -key "certs/servers/localhost/privkey.pem"
+    mkdir -p certs/{servers,tmp}
+    mkdir -p "certs/servers/localhost"
 
-  if test "${cert_domain-set}" == set; then
-    openssl req -key "certs/servers/localhost/privkey.pem" -new -sha256 -out "certs/tmp/localhost.csr.pem" -subj "/C=DE/ST=BW/L=Stuttgart/O=Dummy Authority/CN=localhost"
-  else
-    openssl req -key "certs/servers/localhost/privkey.pem" -new -sha256 -out "certs/tmp/localhost.csr.pem" -subj "/C=DE/ST=BW/L=Stuttgart/O=Dummy Authority/CN=${cert_domain}"
+    openssl genrsa -out "certs/servers/localhost/privkey.pem" 2048 -key "certs/servers/localhost/privkey.pem"
+
+    if test "${cert_domain-set}" == set; then
+      openssl req -key "certs/servers/localhost/privkey.pem" -new -sha256 -out "certs/tmp/localhost.csr.pem" -subj "/C=DE/ST=BW/L=Stuttgart/O=Dummy Authority/CN=localhost"
+    else
+      openssl req -key "certs/servers/localhost/privkey.pem" -new -sha256 -out "certs/tmp/localhost.csr.pem" -subj "/C=DE/ST=BW/L=Stuttgart/O=Dummy Authority/CN=${cert_domain}"
+    fi
+
+    openssl x509 -req -in certs/tmp/localhost.csr.pem -CA certs/ca/root.crt.pem -CAkey certs/ca/root.key.pem -CAcreateserial -out certs/servers/localhost/cert.pem -days 9131
+    mv certs/servers/localhost/privkey.pem /etc/loolwsd/key.pem
+    mv certs/servers/localhost/cert.pem /etc/loolwsd/cert.pem
+    mv certs/ca/root.crt.pem /etc/loolwsd/ca-chain.cert.pem
   fi
 
-  openssl x509 -req -in certs/tmp/localhost.csr.pem -CA certs/ca/root.crt.pem -CAkey certs/ca/root.key.pem -CAcreateserial -out certs/servers/localhost/cert.pem -days 9131
-  mv certs/servers/localhost/privkey.pem /etc/loolwsd/key.pem
-  mv certs/servers/localhost/cert.pem /etc/loolwsd/cert.pem
-  mv certs/ca/root.crt.pem /etc/loolwsd/ca-chain.cert.pem
 fi
 
 # Run loolwsd when asked to do so
